Remove `dojos` word from context name and use 2 previews folders
The dojo description was based on the default location inside `dojos`
folder and showed only the language. With this, the 2 last folder levels
are used. This makes the default description show the dojo name together
with the language and on different locations the base dojo folder and
its parent are used.

---

Changelog: feature