Fix dependency on problem folder re-generating setup every time
This is due to the fact that file generation depends on `${folder}`
existence and make check the timestamp for folder update. As many files
are generated into the folder, that by itself update the timestamp
again and invalidates the timestamp check.

The solution is to avoid file generation inside `${folder}` to
explicitly depend on it.
---

Github-issue: #7
Github-action: fix #7
Changelog: fix